requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alabaster==0.7.13
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
Babel==2.12.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
chex==0.1.7
click==8.1.3
datasets==2.11.0
dgl==1.0.1+cu113
dglgo==0.0.2
dill==0.3.6
dm-tree==0.1.8
docutils==0.19
evaluate==0.4.0
filelock==3.10.7
frozenlist==1.3.3
fsspec==2023.3.0
google-auth==2.16.3
google-auth-oauthlib==0.4.6
grpcio==1.53.0
huggingface-hub==0.13.3
idna==3.4
imagesize==1.4.1
importlib-metadata==6.1.0
isort==5.12.0
jax==0.4.8
jaxlib==0.4.7
Jinja2==3.1.2
joblib==1.2.0
littleutils==0.2.2
Markdown==3.4.3
MarkupSafe==2.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
ml-dtypes==0.0.4
multidict==6.0.4
multiprocess==0.70.14
networkx==3.0
nltk==3.7
numpy==1.19.5
oauthlib==3.2.2
ogb==1.3.5
opt-einsum==3.3.0
optax==0.1.4
outdated==0.2.2
packaging==23.0
pandas==1.5.3
Pillow==9.0.0
protobuf==4.22.1
psutil==5.9.4
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pydantic==1.10.7
pyDeprecate==0.3.2
Pygments==2.14.0
pytorch-lightning==1.7.7
pytz==2023.2
PyYAML==6.0
rdkit-pypi==2022.9.5
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rouge==1.0.1
rouge-score==0.1.2
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sacremoses==0.0.53
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.97
six==1.16.0
snowballstemmer==2.2.0
Sphinx==6.1.3
tensorboard==2.12.0
tokenizers==0.10.3
tomli==2.0.1
toolz==0.12.0
torch-summary==1.4.5
tqdm==4.65.0
transformers==4.8.2
typer==0.7.0
typing_extensions==3.10.0.2
urllib3==1.26.15
Werkzeug==2.2.3
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
